+++ src/Client.php
@@ -8,7 +8,6 @@ use GuzzleHttp\Promise\PromiseInterface;
 use Jean85\PrettyVersions;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
-use Sentry\Exception\EventCreationException;
 use Sentry\Integration\IntegrationInterface;
 use Sentry\Integration\IntegrationRegistry;
 use Sentry\Serializer\RepresentationSerializer;
@@ -126,12 +125,11 @@ final class Client implements ClientInterface
      */
     public function captureMessage(string $message, ?Severity $level = null, ?Scope $scope = null): ?EventId
     {
-        $payload = [
-            'message' => $message,
-            'level' => $level,
-        ];
+        $event = Event::createEvent();
+        $event->setMessage($message);
+        $event->setLevel($level);
 
-        return $this->captureEvent($payload, $scope);
+        return $this->captureEvent($event, null, $scope);
     }
 
     /**
@@ -139,15 +137,17 @@ final class Client implements ClientInterface
      */
     public function captureException(\Throwable $exception, ?Scope $scope = null): ?EventId
     {
-        return $this->captureEvent(['exception' => $exception], $scope);
+        return $this->captureEvent(Event::createEvent(), EventHint::fromArray([
+            'exception' => $exception,
+        ]), $scope);
     }
 
     /**
      * {@inheritdoc}
      */
-    public function captureEvent($payload, ?Scope $scope = null): ?EventId
+    public function captureEvent(Event $event, ?EventHint $hint = null, ?Scope $scope = null): ?EventId
     {
-        $event = $this->prepareEvent($payload, $scope);
+        $event = $this->prepareEvent($event, $hint, $scope);
 
         if (null === $event) {
             return null;
@@ -205,18 +205,32 @@ final class Client implements ClientInterface
     /**
      * Assembles an event and prepares it to be sent of to Sentry.
      *
-     * @param array<string, mixed>|Event $payload The payload that will be converted to an Event
-     * @param Scope|null                 $scope   Optional scope which enriches the Event
+     * @param Event          $event The payload that will be converted to an Event
+     * @param EventHint|null $hint  May contain additional information about the event
+     * @param Scope|null     $scope Optional scope which enriches the Event
      *
      * @return Event|null The prepared event object or null if it must be discarded
      */
-    private function prepareEvent($payload, ?Scope $scope = null): ?Event
+    private function prepareEvent(Event $event, ?EventHint $hint = null, ?Scope $scope = null): ?Event
     {
-        if ($this->options->shouldAttachStacktrace() && !($payload instanceof Event) && !isset($payload['exception']) && !isset($payload['stacktrace'])) {
-            $event = $this->buildEventWithStacktrace($payload);
-        } else {
-            $event = $this->buildEvent($payload);
+        if (null !== $hint) {
+            if (null !== $hint->exception && empty($event->getExceptions())) {
+                $this->addThrowableToEvent($event, $hint->exception);
+            }
+
+            if (null !== $hint->stacktrace && null === $event->getStacktrace()) {
+                $event->setStacktrace($hint->stacktrace);
+            }
         }
+
+        $this->addMissingStacktraceToEvent($event);
+
+        $event->setSdkIdentifier($this->sdkIdentifier);
+        $event->setSdkVersion($this->sdkVersion);
+        $event->setServerName($this->options->getServerName());
+        $event->setRelease($this->options->getRelease());
+        $event->setTags($this->options->getTags());
+        $event->setEnvironment($this->options->getEnvironment());
 
         $sampleRate = $this->options->getSampleRate();
 
@@ -228,7 +242,7 @@ final class Client implements ClientInterface
 
         if (null !== $scope) {
             $previousEvent = $event;
-            $event = $scope->applyToEvent($event, $payload);
+            $event = $scope->applyToEvent($event, $hint);
 
             if (null === $event) {
                 $this->logger->info('The event will be discarded because one of the event processors returned "null".', ['event' => $previousEvent]);
@@ -248,81 +262,33 @@ final class Client implements ClientInterface
     }
 
     /**
-     * Create an {@see Event} with a stacktrace attached to it.
+     * Optionally adds a missing stacktrace to the Event if the client is configured to do so.
      *
-     * @param array<string, mixed> $payload The data to be attached to the Event
+     * @param Event $event The Event to add the missing stacktrace to
      */
-    private function buildEventWithStacktrace($payload): Event
+    private function addMissingStacktraceToEvent(Event $event): void
     {
-        if (!isset($payload['stacktrace']) || !$payload['stacktrace'] instanceof Stacktrace) {
-            $payload['stacktrace'] = $this->stacktraceBuilder->buildFromBacktrace(
-                debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS),
-                __FILE__,
-                __LINE__ - 3
-            );
+        if (!$this->options->shouldAttachStacktrace()) {
+            return;
         }
 
-        return $this->buildEvent($payload);
-    }
-
-    /**
-     * Create an {@see Event} from a data payload.
-     *
-     * @param array<string, mixed>|Event $payload The data to be attached to the Event
-     */
-    private function buildEvent($payload): Event
-    {
-        try {
-            if ($payload instanceof Event) {
-                $event = $payload;
-            } else {
-                $event = Event::createEvent();
-
-                if (isset($payload['logger'])) {
-                    $event->setLogger($payload['logger']);
-                }
-
-                $message = isset($payload['message']) ? mb_substr($payload['message'], 0, $this->options->getMaxValueLength()) : null;
-                $messageParams = $payload['message_params'] ?? [];
-                $messageFormatted = isset($payload['message_formatted']) ? mb_substr($payload['message_formatted'], 0, $this->options->getMaxValueLength()) : null;
-
-                if (null !== $message) {
-                    $event->setMessage($message, $messageParams, $messageFormatted);
-                }
-
-                if (isset($payload['exception']) && $payload['exception'] instanceof \Throwable) {
-                    $this->addThrowableToEvent($event, $payload['exception']);
-                }
-
-                if (isset($payload['level']) && $payload['level'] instanceof Severity) {
-                    $event->setLevel($payload['level']);
-                }
-
-                if (isset($payload['stacktrace']) && $payload['stacktrace'] instanceof Stacktrace) {
-                    $event->setStacktrace($payload['stacktrace']);
-                }
-            }
-        } catch (\Throwable $exception) {
-            throw new EventCreationException($exception);
+        // We should not add a stacktrace when the event already has one or contains exceptions
+        if (null !== $event->getStacktrace() || !empty($event->getExceptions())) {
+            return;
         }
 
-        $event->setSdkIdentifier($this->sdkIdentifier);
-        $event->setSdkVersion($this->sdkVersion);
-        $event->setServerName($this->options->getServerName());
-        $event->setRelease($this->options->getRelease());
-        $event->setTags($this->options->getTags());
-        $event->setEnvironment($this->options->getEnvironment());
-
-        return $event;
+        $event->setStacktrace($this->stacktraceBuilder->buildFromBacktrace(
+            debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS),
+            __FILE__,
+            __LINE__ - 3
+        ));
     }
 
     /**
      * Stores the given exception in the passed event.
      *
-     * @param Event      $event     The event that will be enriched with the
-     *                              exception
-     * @param \Throwable $exception The exception that will be processed and
-     *                              added to the event
+     * @param Event      $event     The event that will be enriched with the exception
+     * @param \Throwable $exception The exception that will be processed and added to the event
      */
     private function addThrowableToEvent(Event $event, \Throwable $exception): void
     {