+++ tests/State/ScopeTest.php
@@ -7,6 +7,7 @@ namespace Sentry\Tests\State;
 use PHPUnit\Framework\TestCase;
 use Sentry\Breadcrumb;
 use Sentry\Event;
+use Sentry\EventHint;
 use Sentry\Severity;
 use Sentry\State\Scope;
 use Sentry\UserDataBag;
@@ -16,7 +17,7 @@ final class ScopeTest extends TestCase
     public function testSetTag(): void
     {
         $scope = new Scope();
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertEmpty($event->getTags());
@@ -24,7 +25,7 @@ final class ScopeTest extends TestCase
         $scope->setTag('foo', 'bar');
         $scope->setTag('bar', 'baz');
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame(['foo' => 'bar', 'bar' => 'baz'], $event->getTags());
@@ -35,14 +36,14 @@ final class ScopeTest extends TestCase
         $scope = new Scope();
         $scope->setTags(['foo' => 'bar']);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame(['foo' => 'bar'], $event->getTags());
 
         $scope->setTags(['bar' => 'baz']);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame(['foo' => 'bar', 'bar' => 'baz'], $event->getTags());
@@ -53,14 +54,14 @@ final class ScopeTest extends TestCase
         $scope = new Scope();
         $scope->setContext('foo', ['foo' => 'bar']);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame(['foo' => ['foo' => 'bar']], $event->getContexts());
 
         $scope->removeContext('foo');
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame([], $event->getContexts());
@@ -69,7 +70,7 @@ final class ScopeTest extends TestCase
     public function testSetExtra(): void
     {
         $scope = new Scope();
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertEmpty($event->getExtra());
@@ -77,7 +78,7 @@ final class ScopeTest extends TestCase
         $scope->setExtra('foo', 'bar');
         $scope->setExtra('bar', 'baz');
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame(['foo' => 'bar', 'bar' => 'baz'], $event->getExtra());
@@ -88,14 +89,14 @@ final class ScopeTest extends TestCase
         $scope = new Scope();
         $scope->setExtras(['foo' => 'bar']);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame(['foo' => 'bar'], $event->getExtra());
 
         $scope->setExtras(['bar' => 'baz']);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame(['foo' => 'bar', 'bar' => 'baz'], $event->getExtra());
@@ -104,7 +105,7 @@ final class ScopeTest extends TestCase
     public function testSetUser(): void
     {
         $scope = new Scope();
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertNull($event->getUser());
@@ -114,7 +115,7 @@ final class ScopeTest extends TestCase
 
         $scope->setUser($user);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame($user, $event->getUser());
@@ -125,7 +126,7 @@ final class ScopeTest extends TestCase
 
         $scope->setUser(['ip_address' => '127.0.0.1', 'subscription_expires_at' => '2020-08-26']);
 
-        $event = $scope->applyToEvent($event, []);
+        $event = $scope->applyToEvent($event);
 
         $this->assertNotNull($event);
         $this->assertEquals($user, $event->getUser());
@@ -145,14 +146,14 @@ final class ScopeTest extends TestCase
         $scope = new Scope();
         $scope->setUser(UserDataBag::createFromUserIdentifier('unique_id'));
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertNotNull($event->getUser());
 
         $scope->removeUser();
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertNull($event->getUser());
@@ -161,14 +162,14 @@ final class ScopeTest extends TestCase
     public function testSetFingerprint(): void
     {
         $scope = new Scope();
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertEmpty($event->getFingerprint());
 
         $scope->setFingerprint(['foo', 'bar']);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame(['foo', 'bar'], $event->getFingerprint());
@@ -177,14 +178,14 @@ final class ScopeTest extends TestCase
     public function testSetLevel(): void
     {
         $scope = new Scope();
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertNull($event->getLevel());
 
         $scope->setLevel(Severity::debug());
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertEquals(Severity::debug(), $event->getLevel());
@@ -197,7 +198,7 @@ final class ScopeTest extends TestCase
         $breadcrumb2 = new Breadcrumb(Breadcrumb::LEVEL_ERROR, Breadcrumb::TYPE_ERROR, 'error_reporting');
         $breadcrumb3 = new Breadcrumb(Breadcrumb::LEVEL_ERROR, Breadcrumb::TYPE_ERROR, 'error_reporting');
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertEmpty($event->getBreadcrumbs());
@@ -205,14 +206,14 @@ final class ScopeTest extends TestCase
         $scope->addBreadcrumb($breadcrumb1);
         $scope->addBreadcrumb($breadcrumb2);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame([$breadcrumb1, $breadcrumb2], $event->getBreadcrumbs());
 
         $scope->addBreadcrumb($breadcrumb3, 2);
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertSame([$breadcrumb2, $breadcrumb3], $event->getBreadcrumbs());
@@ -225,14 +226,14 @@ final class ScopeTest extends TestCase
         $scope->addBreadcrumb(new Breadcrumb(Breadcrumb::LEVEL_ERROR, Breadcrumb::TYPE_ERROR, 'error_reporting'));
         $scope->addBreadcrumb(new Breadcrumb(Breadcrumb::LEVEL_ERROR, Breadcrumb::TYPE_ERROR, 'error_reporting'));
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertNotEmpty($event->getBreadcrumbs());
 
         $scope->clearBreadcrumbs();
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertEmpty($event->getBreadcrumbs());
@@ -256,7 +257,7 @@ final class ScopeTest extends TestCase
             return $eventArg;
         });
 
-        $this->assertSame($event, $scope->applyToEvent($event, []));
+        $this->assertSame($event, $scope->applyToEvent($event));
         $this->assertTrue($callback1Called);
 
         $scope->addEventProcessor(function () use ($callback1Called, &$callback2Called, $callback3Called) {
@@ -274,9 +275,30 @@ final class ScopeTest extends TestCase
             return null;
         });
 
-        $this->assertNull($scope->applyToEvent($event, []));
+        $this->assertNull($scope->applyToEvent($event));
         $this->assertTrue($callback2Called);
         $this->assertFalse($callback3Called);
+    }
+
+    public function testEventProcessorReceivesTheEventAndEventHint(): void
+    {
+        $event = Event::createEvent();
+        $scope = new Scope();
+        $hint = new EventHint();
+
+        $processorCalled = false;
+        $processorReceivedHint = null;
+
+        $scope->addEventProcessor(function (Event $eventArg, EventHint $hint) use (&$processorCalled, &$processorReceivedHint): ?Event {
+            $processorCalled = true;
+            $processorReceivedHint = $hint;
+
+            return $eventArg;
+        });
+
+        $this->assertSame($event, $scope->applyToEvent($event, $hint));
+        $this->assertSame($hint, $processorReceivedHint);
+        $this->assertTrue($processorCalled);
     }
 
     public function testClear(): void
@@ -292,7 +314,7 @@ final class ScopeTest extends TestCase
         $scope->setUser(UserDataBag::createFromUserIdentifier('unique_id'));
         $scope->clear();
 
-        $event = $scope->applyToEvent(Event::createEvent(), []);
+        $event = $scope->applyToEvent(Event::createEvent());
 
         $this->assertNotNull($event);
         $this->assertNull($event->getLevel());
@@ -321,7 +343,7 @@ final class ScopeTest extends TestCase
         $scope->setContext('foocontext', ['foo' => 'bar']);
         $scope->setContext('barcontext', ['bar' => 'foo']);
 
-        $this->assertSame($event, $scope->applyToEvent($event, []));
+        $this->assertSame($event, $scope->applyToEvent($event));
         $this->assertTrue($event->getLevel()->isEqualTo(Severity::warning()));
         $this->assertSame(['foo'], $event->getFingerprint());
         $this->assertSame([$breadcrumb], $event->getBreadcrumbs());